add cliopatria script
