+++ seshat/apps/core/management/commands/populate_videodata.py
@@ -1,7 +1,5 @@
 import os
 import json
-import geopandas as gpd
-from distinctipy import get_colors, get_hex
 from django.contrib.gis.geos import GEOSGeometry, MultiPolygon
 from django.core.management.base import BaseCommand
 from django.db import connection
@@ -16,6 +14,7 @@ class Command(BaseCommand):
 
     def handle(self, *args, **options):
 
+        # TODO: change this to use JSON instead of GeoPandas
         # Load the Cliopatria shape dataset with GeoPandas
         cliopatria_geojson_path = options['geojson_file']
         self.stdout.write(self.style.SUCCESS(f"Loading Cliopatria shape dataset from {cliopatria_geojson_path}..."))
@@ -77,55 +76,3 @@ class Command(BaseCommand):
             """)
         self.stdout.write(self.style.SUCCESS('Simplified geometries added'))
 
-
-def cliopatria_gdf(gdf):
-    """
-    Load the Cliopatria shape dataset with GeoPandas, process names and colors efficiently.
-    """
-
-    # Remove parentheses from 'Name' and 'MemberOf'
-    gdf['CleanName'] = gdf['Name'].str.replace('[()]', '', regex=True)
-    gdf['CleanMember_of'] = gdf['MemberOf'].str.replace('[()]', '', regex=True)
-
-    # Initialize DisplayName and ColorKey
-    gdf['DisplayName'] = gdf['CleanName']
-    gdf['ColorKey'] = gdf['CleanName']
-
-    # Conditions for setting DisplayName to None: If the shape has components, is not a personal union, and the components don't have components
-    has_components = gdf['Components'].notna() & gdf['Components'].str.len() > 0
-    not_personal_union = ~gdf['SeshatID'].str.contains(';')
-    components_without_components = ~gdf['Components'].str.contains('\\(')
-    gdf.loc[has_components & not_personal_union & components_without_components, 'DisplayName'] = None
-
-    # Correct Update ColorKey for shapes that are components of another shape
-    gdf.loc[gdf['MemberOf'].notna() & gdf['MemberOf'].str.len() > 0, 'ColorKey'] = gdf['CleanMember_of']
-
-    # Add type prefix to DisplayName where type is not 'POLITY'
-    gdf.loc[gdf['Type'] != 'POLITY', 'DisplayName'] = gdf['Type'] + ': ' + gdf['DisplayName']
-
-    self.stdout.write(self.style.SUCCESS(f"Generated shape names for {len(gdf)} shapes."))
-    self.stdout.write(self.style.SUCCESS(f"Assigning colours to shapes..."))
-
-    # Use DistinctiPy package to assign a colour based on the ColorKey field
-    colour_keys = gdf['ColorKey'].unique()
-    colours = [get_hex(col) for col in get_colors(len(colour_keys))]
-    colour_mapping = dict(zip(colour_keys, colours))
-
-    # Map colors to a new column efficiently
-    gdf['Color'] = gdf['ColorKey'].map(colour_mapping)
-
-    # Drop intermediate columns
-    gdf.drop(['CleanName', 'CleanMember_of'], axis=1, inplace=True)
-
-    self.stdout.write(self.style.SUCCESS(f"Assigned colours to {len(gdf)} shapes."))
-    self.stdout.write(self.style.SUCCESS(f"Determing polity start and end years..."))
-
-    # Add a column called 'PolityStartYear' to the GeoDataFrame which is the minimum 'FromYear' of all shapes with the same 'Name'
-    gdf['PolityStartYear'] = gdf.groupby('Name')['FromYear'].transform('min')
-
-    # Add a column called 'PolityEndYear' to the GeoDataFrame which is the maximum 'ToYear' of all shapes with the same 'Name'
-    gdf['PolityEndYear'] = gdf.groupby('Name')['ToYear'].transform('max')
-
-    self.stdout.write(self.style.SUCCESS(f"Determined polity start and end years for {len(gdf)} shapes."))
-
-    return gdf
